wntd/preonboarding/requirements.txt
django~=4.1
djangorestframework~=3.13.1
